+++ modules/apps/27-interchain-accounts/types/expected_keepers.go
@@ -37,5 +37,7 @@ type PortKeeper interface {
 // ParamSubspace defines the expected Subspace interface for module parameters.
 type ParamSubspace interface {
 	GetParamSet(ctx sdk.Context, ps paramtypes.ParamSet)
-	GetParamSetIfExists(ctx sdk.Context, ps paramtypes.ParamSet)
+	Has(ctx sdk.Context, key []byte) bool
+	Get(ctx sdk.Context, key []byte, ptr interface{})
+	Set(ctx sdk.Context, key []byte, value interface{})
 }